created create event interactions handler to separate concerns
