+++ app/src/main/java/com/example/c4q/capstone/userinterface/events/CreateEventPresenter.java
@@ -13,6 +13,7 @@ import java.util.List;
 
 /**
  * Created by amirahoxendine on 3/19/18.
+ * This is a class to handle data input for CreateEventFragment
  */
 
 public class CreateEventPresenter {
