+++ app/src/main/res/layout/fragment_create_event.xml
@@ -113,11 +113,11 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginVertical="@dimen/layout_item_margin"
-        android:orientation="vertical">
+        android:orientation="horizontal">
 
         <Button
             android:id="@+id/add_group_button"
-            android:layout_width="match_parent"
+            android:layout_width="0dp"
             android:layout_height="wrap_content"
             android:layout_margin="@dimen/layout_item_margin"
             android:layout_weight="1"
@@ -125,30 +125,9 @@
             android:text="@string/add_group"
             android:textColor="@color/dark_gray"
             android:tag="add group button"/>
-        <FrameLayout
-            android:id="@+id/add_group_fragment_container"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:visibility="gone">
-            <TextView
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:text="Group Fragment goes here"
-                android:gravity="center"
-                android:textAppearance="@style/TextAppearance.AppCompat.Title"/>
-        </FrameLayout>
-
-    </LinearLayout>
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginVertical="@dimen/layout_item_margin"
-        android:orientation="vertical">
-
         <Button
             android:id="@+id/add_friends_button"
-            android:layout_width="match_parent"
+            android:layout_width="0dp"
             android:layout_height="wrap_content"
             android:layout_margin="@dimen/layout_item_margin"
             android:layout_weight="1"
@@ -156,20 +135,24 @@
             android:text="@string/add_friends"
             android:textColor="@color/dark_gray"
             android:tag="add friends button"/>
-        <FrameLayout
-            android:id="@+id/add_friends_fragment_container"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:visibility="gone">
-            <TextView
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:text="Friends Fragment goes here"
-                android:gravity="center"
-                android:textAppearance="@style/TextAppearance.AppCompat.Title"/>
-            </FrameLayout>
     </LinearLayout>
 
+    <FrameLayout
+        android:id="@+id/invite_guests_fragment_container"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:visibility="gone">
+        <TextView
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="Invite guests gragment  goes here"
+            android:gravity="center"
+            android:textAppearance="@style/TextAppearance.AppCompat.Title"/>
+    </FrameLayout>
+    <TextView
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="@string/add_note"/>
     <EditText
         android:id="@+id/add_note"
         android:layout_width="match_parent"
